network: Determine if netns was provided or created
This commit main purpose is to provide an indication about the netns
so that we can know if it was created internally or if we got it from
the network configuration provided.

The second purpose of this commit is to clean and re-organize the way
the config was handled. A config is not suppposed to be modified, it
only provides inputs.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>