hyperstart: Limit the hostname length to prevent from hyperstart errors
Using the pod ID as the hostname was not a bad idea but it can
sometimes be longer than 64 bytes. In that case, hyperstart returns
an error because it cannot properly set the hostname. This is a Linux
limitation related to sethostname() system call. It only accepts names
not longer than 64 bytes.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>